# openapi/swagger.yaml
openapi: 3.0.1
info:
  title: Process management
  description: Process management
  version: 1.0.0
servers:
- url: /
paths:
...
  /vehicleStatus/{id}:
    get:
      tags:
        - VehicleStatus
      operationId: getVehicleStatus
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: efeu ID
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleStatus'
    post:
      tags:
        - VehicleStatus
      operationId: createVehicleStatus
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: efeu ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleStatus'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleStatus'
    put:
      tags:
        - VehicleStatus
      operationId: updateVehicleStatus
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: efeu ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleStatus'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleStatus'
    delete:
      tags:
        - VehicleStatus
      operationId: deleteVehicleStatus
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: efeu ID
      responses:
        200:
          description: Success
          content:
            text/plain:
              schema:
                type: string
...
components:
  schemas:
...
    VehicleStatus:
      required:
          - id
          - state
          - latitude
          - longitude
          - remainingRange
          - timestamp
      type: object
      properties:
        id:
          type: string
        vehicleState:
          type: string
          enum:
            - Idle
            - Driving
            - Recharging
            - Loading
            - Waiting
            - Disrupted
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        boxId:
          type: string
        remainingRange:
          type: integer
        timestamp:
          type: string
          format: date-time
...