# src/main/java/de/fzi/efeu/api/VehicleStatusApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (4.3.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package de.fzi.efeu.api;

import java.util.List;
import java.util.Optional;

import javax.validation.Valid;

import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.context.request.NativeWebRequest;

import de.fzi.efeu.model.VehicleStatus;
import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2020-07-10T14:10:36.203026900+02:00[Europe/Berlin]")

@Validated
@Api(value = "vehicleStatus", description = "the vehicleStatus API")
public interface VehicleStatusApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /vehicleStatus/{id}
     *
     * @param id efeu ID (required)
     * @param vehicleStatus  (optional)
     * @return Success (status code 200)
     */
    @ApiOperation(value = "", nickname = "createVehicleStatus", notes = "", response = VehicleStatus.class, tags={ "VehicleStatus", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Success", response = VehicleStatus.class) })
    @RequestMapping(value = "/vehicleStatus/{id}",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    default ResponseEntity<VehicleStatus> createVehicleStatus(
            @ApiParam(value = "efeu ID", required = true) @PathVariable("id") String id,
            @ApiParam(value = "") @Valid @RequestBody(required = false) VehicleStatus vehicleStatus) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"remainingRange\" : 1, \"latitude\" : 0.8008282, \"id\" : \"id\", \"vehicleState\" : \"Idle\", \"longitude\" : 6.0274563, \"boxId\" : \"boxId\", \"timestamp\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /vehicleStatus/{id}
     *
     * @param id efeu ID (required)
     * @return Success (status code 200)
     */
    @ApiOperation(value = "", nickname = "deleteVehicleStatus", notes = "", response = String.class, tags={ "VehicleStatus", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Success", response = String.class) })
    @RequestMapping(value = "/vehicleStatus/{id}",
        produces = { "text/plain" }, 
        method = RequestMethod.DELETE)
    default ResponseEntity<String> deleteVehicleStatus(
            @ApiParam(value = "efeu ID", required = true) @PathVariable("id") String id) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /vehicleStatus/{id}
     *
     * @param id efeu ID (required)
     * @return Success (status code 200)
     */
    @ApiOperation(value = "", nickname = "getVehicleStatus", notes = "", response = VehicleStatus.class, tags={ "VehicleStatus", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Success", response = VehicleStatus.class) })
    @RequestMapping(value = "/vehicleStatus/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<VehicleStatus> getVehicleStatus(
            @ApiParam(value = "efeu ID", required = true) @PathVariable("id") String id) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"remainingRange\" : 1, \"latitude\" : 0.8008282, \"id\" : \"id\", \"vehicleState\" : \"Idle\", \"longitude\" : 6.0274563, \"boxId\" : \"boxId\", \"timestamp\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /vehicleStatus/
     *
     * @return Success (status code 200)
     */
    @ApiOperation(value = "", nickname = "getVehicleStatusList", notes = "", response = VehicleStatus.class, responseContainer = "List", tags={ "VehicleStatus", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Success", response = VehicleStatus.class, responseContainer = "List") })
    @RequestMapping(value = "/vehicleStatus/",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<List<VehicleStatus>> getVehicleStatusList() {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"remainingRange\" : 1, \"latitude\" : 0.8008282, \"id\" : \"id\", \"vehicleState\" : \"Idle\", \"longitude\" : 6.0274563, \"boxId\" : \"boxId\", \"timestamp\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /vehicleStatus/{id}
     *
     * @param id efeu ID (required)
     * @param vehicleStatus  (optional)
     * @return Success (status code 200)
     */
    @ApiOperation(value = "", nickname = "updateVehicleStatus", notes = "", response = VehicleStatus.class, tags={ "VehicleStatus", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Success", response = VehicleStatus.class) })
    @RequestMapping(value = "/vehicleStatus/{id}",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.PUT)
    default ResponseEntity<VehicleStatus> updateVehicleStatus(
            @ApiParam(value = "efeu ID", required = true) @PathVariable("id") String id,
            @ApiParam(value = "") @Valid @RequestBody(required = false) VehicleStatus vehicleStatus) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"remainingRange\" : 1, \"latitude\" : 0.8008282, \"id\" : \"id\", \"vehicleState\" : \"Idle\", \"longitude\" : 6.0274563, \"boxId\" : \"boxId\", \"timestamp\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}